Move and further cleaning
